Refactor Clock component
