+++ src/hooks/clock/index.tsx
@@ -1,0 +1,33 @@
+import { useEffect, useState } from "react";
+import { useDashboardSettings } from "../../context/DashboardSettingsContext";
+
+const getDateTime = (clockFormat: "12" | "24") => {
+  const now = new Date();
+
+  return {
+    currentDate: now
+      .toLocaleString("en-US", { dateStyle: "full" })
+      .toUpperCase(),
+    currentTime: now.toLocaleString("en-US", {
+      hour: "2-digit",
+      hour12: clockFormat === "12",
+      minute: "2-digit",
+    }),
+  };
+};
+
+export const useClock = () => {
+  const { clockFormat } = useDashboardSettings();
+
+  const [dateTime, setDateTime] = useState(getDateTime(clockFormat));
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setDateTime(getDateTime(clockFormat));
+    }, 1000);
+
+    return () => clearInterval(interval);
+  }, [clockFormat]);
+
+  return { dateTime };
+};
